+++ src/renderer/render/layers.ts
@@ -1,13 +1,5 @@
 export enum Layers {
-    Player1,
-    Player2,
-    Player3,
-    Player4,
-    Player5,
-    Player6,
-    Player7,
-    Player8,
-    Resources,
-    Doodads,
+    Terrain = 10,
+    Units,
     PictureInPicture
 }