+++ src/renderer/utils/camera-utils.ts
@@ -1,5 +1,5 @@
 import CameraControls from "camera-controls";
-import { Box3, Camera, MathUtils, Vector3, Vector4 } from "three";
+import { Box3, Camera, MathUtils, Object3D, PerspectiveCamera, Vector3, Vector4 } from "three";
 import type { CameraMouse } from "../input/camera-mouse";
 import type { CameraKeys } from "../input/camera-keys";
 import type CameraShake from "../camera/camera-shake";
@@ -84,4 +84,41 @@ export function applyCameraDirectionToImageFrame(camera: DirectionalCamera, imag
     _imageFrameInfo.frame = image.frameIndexBase + newFrameOffset.frame;
   }
   return _imageFrameInfo;
+}
+
+// https://codepen.io/discoverthreejs/pen/vwVeZB
+const _target = new Vector3();
+export const zoomCameraToSelection = (camera: PerspectiveCamera, controls: CameraControls, selection: Object3D[], fitRatio = 1.2) => {
+  const box = new Box3();
+
+  for (const object of selection) box.expandByObject(object);
+
+  const size = box.getSize(new Vector3());
+  const center = box.getCenter(new Vector3());
+
+  const maxSize = Math.max(size.x, size.y, size.z);
+  const fitHeightDistance =
+    maxSize / (2 * Math.atan((Math.PI * camera.fov) / 360));
+  const fitWidthDistance = fitHeightDistance / camera.aspect;
+  const distance = fitRatio * Math.max(fitHeightDistance, fitWidthDistance);
+
+  controls.getTarget(_target)
+  const direction = _target
+    .clone()
+    .sub(camera.position)
+    .normalize()
+    .multiplyScalar(distance);
+
+  // controls.maxDistance = distance * 10;
+  controls.setTarget(center.x, center.y, center.z);
+
+  center.sub(direction);
+  controls.moveTo(center.x, center.y, center.z);
+
+  // camera.near = distance / 100;
+  // camera.far = distance * 100;
+  // camera.updateProjectionMatrix();
+
+  // camera.position.copy(controls.target).sub(direction);
+  return distance;
 }